+++ test/orchard/info_test.clj
@@ -8,6 +8,7 @@
   (is (not-empty (info/see-also 'clojure.core 'map))))
 
 (deftype T [])
+(defrecord R [])
 
 (deftest info-test
   (is (info/info 'orchard.info 'io))
@@ -19,8 +20,8 @@
   (is (info/info 'orchard.info '.toString))
 
   (is (not (info/info 'clojure.core (gensym "non-existing"))))
-  (is (info/info 'orchard.info-test 'T)
-      "Check that deftype T (which returns nil for .getPackage), doesn't throw")
+  (is (info/info 'orchard.info-test 'T))
+  (is (info/info 'orchard.info-test 'R))
 
   (is (= (the-ns 'clojure.core) (:ns (info/info 'orchard.info 'str))))